Generalized to a Pattern base class and added a GradientPattern based on it
Refactored shared logic into a Pattern base class and changed the StripePattern to
inherit from it.
Created a GradientPattern to return points on a gradient from one colour to another.
A complication not handled in the book is that if the range of x-values on the object
covered by the pattern are > 1, then the pattern will cycle through the gradient multiple
times which is a bit icky.  Fixing this requires either customizing the Pattern with
a scaling that reduces the range of x from 0 to 1 (in pattern space), or an extra abstraction
that the book doesn't cover (converting the point of intersection from a 3d point in object
space to a 2d point in pattern space where the co-ordinates are always normalized within the
range [0,1]